requirements.txt
dj-database-url==0.4.1
Django==1.10.3
django-localflavor==1.3
EditorConfig==0.12.1
psycopg2==2.6.2
python-decouple==3.0
python-social-auth==0.2.21
